快猫/requirements.txt
cryptography
pycryptodome
easygui